+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-## 0.2.0
+## 0.2.0 - December 21, 2018
 - Minor bug fixes in STDIO server
 - More efficient client reads
 - Rename server methods `prepare` for clarity
 - Improved socket state handling
 - More accurate interval time
+- Adapter#remote attribute
+- Adapter#close method
 
 ## 0.1.0 - December 20, 2018
 First release